pyproject.toml
[build-system]
requires = ["setuptools>=63", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "algoforge"
version = "0.1.0"
description = "Algorithmic trading engine for stocks and forex"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.2",
    "uvicorn[standard]==0.26.0",
    "gunicorn==21.2.0",
    "pydantic==1.10.14",
    "SQLAlchemy==2.0.25",
    "alembic==1.13.1",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.1",
    "passlib[bcrypt]==1.7.4",
    "python-jose==3.3.0",
    "pyjwt==2.8.0",
    "cryptography==41.0.7",
    "redis==5.0.1",
    "backtrader==1.9.78.123",
    "yfinance==0.2.32",
    "alpha-vantage==2.3.1",
    "pandas==2.1.4",
    "numpy==1.26.3",
    "matplotlib==3.8.2",
    "quantstats==0.0.62",
    "nltk==3.8.1",
    "aiohttp==3.9.1",
    "httpx==0.25.2",
    "itsdangerous==2.1.2",
    "loguru==0.7.2",
    "python-dateutil==2.8.2",
    "tenacity==8.2.3",
    "pycryptodome==3.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.2",
    "pytest-cov==4.1.0",
    "httpx==0.25.2",
    "coverage==7.3.2",
    "mypy==1.8.0",
    "types-redis==4.6.0.10",
    "types-python-dateutil==2.8.19.14",
    "types-PyYAML==6.0.12.12",
    "ruff==0.1.11",
    "black==23.12.1"
]

[tool.setuptools.packages.find]
where = ["app"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["B008"]

[tool.coverage.run]
branch = true
source = ["app"]
